Only enqueue admin scripts and styles when necessary, and add jQuery dependency to JS
The admin CSS and JS for the plugin was being enqueued everywhere on
the administration panels. This is unnecessary, and was causing
problems in some places because jQuery wasn’t added as a dependency of
the JS file (now it is).

This seems to work, but probably deserves more testing.